+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.0 (unreleased)
 
 * ignore output from executor and show only stdout/stderr from actual command (#11)
+* return exit code 1 if at least one task failed (#12)
 
 ## v0.1.0